+++ test-platform-ui/src/app/administration/questions/page.tsx
@@ -1,26 +1,22 @@
 'use client';
 
-import CustomModal, {
-  CustomModalHandler,
-} from '@/components/molecules/CustomModal';
-import DataTable, { multipleLinesTypo } from '@/components/molecules/Grid';
 import { IResponseQuestion } from '@/constants/questions';
 import { ROUTE_KEY } from '@/constants/routePaths';
 import ApiHook, { Methods } from '@/libs/apis/ApiHook';
 import { DataContext } from '@/libs/contextStore';
-import { QuestionLevel } from '@/libs/definitions';
 import { showNotification } from '@/libs/toast';
 import { handleMappingImportData } from '@/libs/utils';
-import { AddBox, Delete, ModeEdit } from '@mui/icons-material';
-import ClearIcon from '@mui/icons-material/Clear';
+import { AddBox } from '@mui/icons-material';
+import AccessTimeIcon from '@mui/icons-material/AccessTime';
 import CloudUploadIcon from '@mui/icons-material/CloudUpload';
-import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
-import { Box, Chip, IconButton, Stack } from '@mui/material';
+import StackedBarChartOutlinedIcon from '@mui/icons-material/StackedBarChartOutlined';
+import TaskAltIcon from '@mui/icons-material/TaskAlt';
+import { Box, Stack } from '@mui/material';
 import Button from '@mui/material/Button';
 import { styled } from '@mui/material/styles';
-import { GridColDef } from '@mui/x-data-grid';
 import { useRouter } from 'next/navigation';
 import React, { useContext, useEffect, useState } from 'react';
+import ListCardItem from './grid-components/listItem';
 
 export interface IQuestion {
   id: number;
@@ -44,11 +40,9 @@ const VisuallyHiddenInput = styled('input')({
 });
 
 const Page = () => {
-  const [questionList, setQuestionList] = useState<IQuestion[]>([]);
-  const [loading, setLoading] = useState<boolean>(true);
-  const onDeleteQuestion = React.useRef<number>(0);
+  const [_questionList, setQuestionList] = useState<IQuestion[]>([]);
+  const [_loading, setLoading] = useState<boolean>(true);
   const router = useRouter();
-  const modalRef = React.useRef<CustomModalHandler>(null);
   const [isImportLoading, setImportLoading] = useState<boolean>(false);
 
   const { data, updateData } = useContext(DataContext);
@@ -78,17 +72,6 @@ const Page = () => {
     getGridQuestion();
   }, []);
 
-  const handleDeleteQuestion = async (questionId: number) => {
-    const { error } = await ApiHook(Methods.DELETE, `/questions/${questionId}`);
-    // Handle response
-    if (!error) {
-      getGridQuestion();
-    } else {
-      alert(`Can not delete question ${questionId}`);
-    }
-    modalRef.current?.close();
-  };
-
   const handleImport = async (event: React.ChangeEvent<HTMLInputElement>) => {
     const files = event.target.files;
     const file = files?.[0];
@@ -113,169 +96,107 @@ const Page = () => {
     };
   };
 
-  const ModalContent = () => {
-    return (
-      <Box className="grid">
-        <h2 id="parent-modal-title">{`Deleting Question ID ${onDeleteQuestion.current}`}</h2>
-        <p id="parent-modal-description" className="mt-3">
-          Are you sure want to delete selected question ? Question will be
-          delete permanently.
-        </p>
-        <Box className="inline-flex gap-10">
-          <Button
-            className="mt-5 w-full content-end"
-            title="Delete"
-            variant="outlined"
-            startIcon={<ClearIcon />}
-            onClick={() => modalRef.current?.close()}
-          >
-            Cancel
-          </Button>
-          <Button
-            className="mt-5 w-full content-end"
-            title="Delete"
-            variant="contained"
-            startIcon={<DeleteForeverIcon />}
-            onClick={() => handleDeleteQuestion(onDeleteQuestion.current)}
-          >
-            Delete
-          </Button>
-        </Box>
-      </Box>
-    );
-  };
-
-  //#region : Temporary definition Questions list's columns:
-  const columns: GridColDef[] = [
-    {
-      field: 'id',
-      headerName: 'ID',
-      disableColumnMenu: true,
-      width: 70,
-    },
-    {
-      field: 'title',
-      headerName: 'Title',
-      flex: 0.6,
-      renderCell: (params) => multipleLinesTypo(params.row.title),
-    },
-    {
-      field: 'level',
-      headerName: 'Standard',
-      flex: 0.2,
-      renderCell: (params) => {
-        const level = QuestionLevel.find(
-          (lvl) => lvl.value === params.row.level,
-        );
-        return (
-          <Box className="grid gap-1">
-            {level ? <Chip label={params.row.level} /> : null}
-          </Box>
-        );
-      },
-    },
-    {
-      field: 'categories',
-      headerName: 'Categories',
-      flex: 0.3,
-      renderCell: (params) => {
-        return (
-          <Box className="grid gap-1">
-            {params.row.categories?.map((cate: any, indx: number) => {
-              return <Chip key={`cate-${indx}`} label={cate} />;
-            })}
-          </Box>
-        );
-      },
-    },
-    {
-      field: 'actions',
-      headerName: 'Actions',
-      sortable: false,
-      flex: 0.2,
-      disableColumnMenu: true,
-      renderCell: (params) => {
-        return (
-          <>
-            <IconButton
-              title="Edit"
-              onClick={() =>
-                router.push(
-                  `${ROUTE_KEY.ADMINISTRATION_QUESTIONS}/${params.row.id}`,
-                )
-              }
-            >
-              <ModeEdit />
-            </IconButton>
-            <IconButton
-              title="Delete"
-              className="ml-3"
-              onClick={() => {
-                onDeleteQuestion.current = params.row.id;
-                modalRef.current?.open();
-              }}
-            >
-              <Delete />
-            </IconButton>
-          </>
-        );
-      },
-    },
-  ];
-  //#endregion
-
   return (
     <Box>
-      <Box className="flex flex-row-reverse items-center justify-between pb-10">
-        <Box>
-          <Button
-            component="label"
-            role={undefined}
-            variant="contained"
-            tabIndex={-1}
-            startIcon={<CloudUploadIcon />}
-            className="mr-3"
-            disabled={isImportLoading}
-          >
-            Import
-            <VisuallyHiddenInput
-              type="file"
-              accept="application/JSON"
-              onChange={handleImport}
-            />
-          </Button>
-          <Button
-            variant="contained"
-            onClick={(evt: React.MouseEvent) => {
-              evt.preventDefault();
-              updateData({
-                ...data,
-                pagination: {
-                  pageNum: 1,
+      <Stack
+        className="h-[calc(100vh_-_120px)]"
+        gridTemplateColumns={'2fr 1fr'}
+        display={'grid'}
+        gap={4}
+      >
+        <Box className="question-list grid h-full gap-3">
+          {[1, 2, 3].map((value: number, _indx: number) => (
+            <ListCardItem
+              cardData={{
+                title: `Render card data title ${value}`,
+                tags: [
+                  'Aggregation',
+                  'Data Filtering',
+                  'Data Manipulation',
+                  'Database Management',
+                ],
+                flagChip: {
+                  label: 'Javascript',
                 },
-              });
-              router.push(ROUTE_KEY.ADMINISTRATION_QUESTIONS_CREATE);
-            }}
-            startIcon={<AddBox />}
-          >
-            New Question
-          </Button>
+                description: {
+                  label: 'Testing description label',
+                  content: `Testing description content. Lorem Ipsum is simply dummy text of the printing and
+                  typesetting industry. Lorem Ipsum has been the standard
+                  dummy text ever since the 1500s`,
+                },
+                cardInfo: {
+                  render: () => {
+                    return (
+                      <>
+                        <span className="info-chip">
+                          <TaskAltIcon fontSize="small" />
+                          Multiple Choice
+                        </span>
+                        <span className="info-chip">
+                          <StackedBarChartOutlinedIcon
+                            sx={{
+                              width: '16px',
+                              height: '16px',
+                            }}
+                          />
+                          Easy
+                        </span>
+                        <span className="info-chip">
+                          <AccessTimeIcon
+                            sx={{
+                              width: '16px',
+                              height: '16px',
+                            }}
+                          />
+                          30 min
+                        </span>
+                      </>
+                    );
+                  },
+                },
+              }}
+            />
+          ))}
         </Box>
-      </Box>
-      <Stack gridTemplateColumns={'2fr 1fr'} display={'grid'} gap={4}>
-        <DataTable
-          className=""
-          rows={questionList}
-          columns={columns}
-          loading={loading}
-          height="h-[calc(100vh_-_215px)]"
-        />
-        <Box className="summary-list border-2 border-dashed border-slate-300"></Box>
+        <Box className="summary-list h-full border-2 border-dashed border-slate-300">
+          <Box className="flex flex-row-reverse items-center justify-between pb-6">
+            <Box>
+              <Button
+                component="label"
+                role={undefined}
+                variant="contained"
+                tabIndex={-1}
+                startIcon={<CloudUploadIcon />}
+                className="mr-3"
+                disabled={isImportLoading}
+              >
+                Import
+                <VisuallyHiddenInput
+                  type="file"
+                  accept="application/JSON"
+                  onChange={handleImport}
+                />
+              </Button>
+              <Button
+                variant="contained"
+                onClick={(evt: React.MouseEvent) => {
+                  evt.preventDefault();
+                  updateData({
+                    ...data,
+                    pagination: {
+                      pageNum: 1,
+                    },
+                  });
+                  router.push(ROUTE_KEY.ADMINISTRATION_QUESTIONS_CREATE);
+                }}
+                startIcon={<AddBox />}
+              >
+                New Question
+              </Button>
+            </Box>
+          </Box>
+        </Box>
       </Stack>
-
-      <CustomModal ref={modalRef} title={''}>
-        <ModalContent />
-      </CustomModal>
     </Box>
   );
 };
